fix: implement actual --dry-run
Potentially we could click the Cancel button, but even better is to just
reload the page. We have to reload anyway after Cancelling. This also
adds `--sleep` as an option. So you can now run:

```
  punch submit --dry-run --headed --sleep 5.0
```

and then see what information would have been submitted, before the next
page is filled out. We could use page.pause() but then the user needs a
way to actually tell it to go to the next, and that probably would
require injecting some sort of item into the page that we could then
wait for.

Also tweak how the messages are updated. Now we write the message about
what item we're working on, but only increment the progress counter once
we've finished filling out that item (rather than only displaying the
item once we've completely submitted it).

This also changes the selectors. Salesforce has accessibility tags, and
when using the built in locator, it seemed to recommend 'get_by_use'
rather than locator with an xpath.